refactor: separate logic into a dedicated file
